Add files via upload
Malaria is a vector-borne infectious disease that affects humans and is caused by the parasites which belong to the plasmodium class: these parasites are called protozoan i.e. a class of one-cell microorganisms). The symptoms of the disease are fatigue, fever, headache, and vomiting in typical cases; and seizures, yellowish skin, coma, and/or death, in severe cases. The symptoms begin to surface ten to fifteen days after an individual is bitten by the Anopheles mosquito (female mosquito) thereby, infecting the person with the parasite which is embedded in its saliva.
About half of the world population (104 countries across the globe) is susceptible to malaria infection, with morbidity of about 241 million cases; 5% incidence, and 12% mortality annually (WHO, 2021). The spread of the disease occurs greatly in both tropical and subtropical regions which include, some parts of Europe, the Middle East, South America, Asia, and most predominantly Africa (especially Sub-Saharan Africa) where most deaths have been reported. It is quite alarming to note that in every 30 seconds, there is a death of one child due to malaria infection in Sub-Saharan Africa – a region that has 90% of all the deaths due to malaria infection. Globally, about 2000 children and youth die daily due to the disease, and also the disease accounts for 30% of hospitalized children (under the age of five) in Nigeria, as well as about 60% of outpatient visits.  
 
The question is should we focus more attention on the treatment of infected individuals or more energy should be channeled towards prevention?